requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.0.4
colorama==0.4.0
cycler==0.10.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
isort==4.3.4
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
matplotlib==3.0.2
mccabe==0.6.1
mock==3.0.5
numpy==1.16.4
pandas==0.23.4
Pillow==5.3.0
protobuf==3.6.1
pylint==2.1.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==5.1.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
